Switch to size_t instead of int where appropriate

+++ source/parlex/include/parlex/builtins/any_character.hpp
@@ -9,8 +9,8 @@ namespace details {
 class any_character_t : public terminal {
 public:
 	virtual ~any_character_t() = default;
-	virtual bool test(std::u32string const & document, int documentPosition) const final;
-	virtual int get_length() const final;
+	virtual bool test(std::u32string const & document, size_t documentPosition) const final;
+	virtual size_t get_length() const final;
 	virtual std::string get_id() const final;
 };
