+++ source/parlex/include/parlex/producer.hpp
@@ -21,10 +21,10 @@ class job;
 class producer {
 public:
 	struct subscription {
-		int next_index;
+		size_t next_index;
 		context_ref const c;
-		int next_dfa_state;
-		inline subscription(context_ref const & c, int const nextDfaState) : next_index(0), c(c), next_dfa_state(nextDfaState) {}
+		size_t next_dfa_state;
+		inline subscription(context_ref const & c, size_t const nextDfaState) : next_index(0), c(c), next_dfa_state(nextDfaState) {}
 	};
 
 	void do_events();
@@ -40,9 +40,9 @@ public:
 
 	friend class parlex::parser;
 	friend class parlex::details::job;
-	void add_subscription(context_ref const & c, int const nextDfaState);
-	producer(job & owner, recognizer const & r, int const documentPosition);
-	void enque_permutation(int consumedCharacterCount, permutation const & p);
+	void add_subscription(context_ref const & c, size_t const nextDfaState);
+	producer(job & owner, recognizer const & r, size_t const documentPosition);
+	void enque_permutation(size_t consumedCharacterCount, permutation const & p);
 };
 
 }
