+++ source/parlex/include/parlex/terminal.hpp
@@ -15,8 +15,8 @@ class terminal : public recognizer {
 public:
 	virtual ~terminal() = default;
 
-	virtual bool test(std::u32string const & document, int documentPosition) const = 0;
-	virtual int get_length() const = 0;
+	virtual bool test(std::u32string const & document, size_t documentPosition) const = 0;
+	virtual size_t get_length() const = 0;
 	inline bool is_terminal() const final { return true; }
 protected:
 	inline terminal() {}
