+++ source/parlex/src/match_class.cpp
@@ -2,7 +2,7 @@
 
 namespace parlex {
 
-match_class::match_class(recognizer const & r, int documentPosition) : r(r), document_position(documentPosition) {}
+match_class::match_class(recognizer const & r, size_t documentPosition) : r(r), document_position(documentPosition) {}
 
 bool match_class::operator <(match_class const & rhs) const {
 	return document_position < rhs.document_position ||
