+++ source/parlex/src/token.cpp
@@ -6,7 +6,7 @@
 namespace parlex {
 namespace details {
 
-token::token(job & owner, terminal const & t, int documentPosition) : producer(owner, t, documentPosition) { 
+token::token(job & owner, terminal const & t, size_t documentPosition) : producer(owner, t, documentPosition) { 
 	if (t.test(owner.document, documentPosition)) {
 		DBG("found a '", t.get_id(), "' at ", documentPosition);
 		enque_permutation(t.get_length(), std::vector<match>());
